Move GenericDocument, DocumentNature, LegalInstrument and Legislation models into peachjam
